Este es el commit con lo de twilio y SendGrid completo, junto con lo de superheroes todo

+++ src/services/superhero_v2.service.js
@@ -1,4 +1,5 @@
-const superheroModel = require('../models/superhero_v2.model')
+const superheroModel = require('../models/superhero_v2.model');
+const Boom=require('@hapi/boom');
 class SuperheroService {
   /* Promesas y funciones asincronicas
     una funcion asincronica devuelve una promesa
@@ -11,18 +12,29 @@ class SuperheroService {
   async listSuperhero() {
     return superheroModel.find();
   }
-  async ShowSuperhero(superheroId) {
-    return superheroModel.findById({
-      _id: superheroId
-    });
 
+  find(){
+    return new Promise((resolve)=>{
+      setTimeout(()=>{
+        resolve(superheroModel.find());
+      },3000);
+    });
+  }
+
+  async showSuperhero(superheroId) {
+    return superheroModel.findById({_id:superheroId}).then(
+      (superheroFind)=>{
+        if (!superheroFind) throw Boom.notFound('no se encontró el super heroe');
+        return superheroFind;
+      }
+    )
   }
   async editSuperhero(superheroId, superhero, realname, superpower) {
     return superheroModel.findById({
       _id: superheroId
     }).then(
       (superheroFind) => {
-        if (!superheroFind) throw Error('no se encontró el super heroe');
+        if (!superheroFind) throw Boom.notFound('no se encontró el super heroe');
         return superheroModel.updateOne({
           superheroId
         }, {
@@ -37,6 +49,7 @@ class SuperheroService {
     const superhero_remove = superheroModel.findById({
       _id: superheroId
     });
+    if(!superheroId) throw Boom.notFound('no se encontró el super heroe');
     return superheroModel.deleteOne(superhero_remove);
   }
